+++ .eslintrc.json
@@ -1,3 +1,7 @@
 {
-  "extends": "airbnb"
+  "extends": "airbnb",
+  "rules": {
+    "react/jsx-filename-extension": "off",
+    "jsx-a11y/label-has-for": "off"
+  }
 }
