+++ package.json
@@ -11,11 +11,14 @@
     "express": "^4.16.3",
     "mapbox-gl": "^0.45.0",
     "migrate-mongo": "^3.0.2",
+    "moment": "^2.22.2",
     "mongo-migrate": "^2.0.2",
     "mongoose": "^5.1.3",
     "nodemon": "^1.17.5",
     "react": "^16.4.0",
+    "react-datetime-picker": "^1.3.1",
     "react-dom": "^16.4.0",
+    "react-moment-datetimepicker": "^0.7.0",
     "react-router-dom": "^4.2.2",
     "react-scripts": "1.1.4"
   },
@@ -28,6 +31,7 @@
   "devDependencies": {
     "eslint": "^4.19.1",
     "eslint-config-airbnb": "^16.1.0",
+    "eslint-config-airbnb-base": "^12.1.0",
     "eslint-plugin-import": "^2.12.0",
     "eslint-plugin-jsx-a11y": "^6.0.3",
     "eslint-plugin-react": "^7.9.1"
