+++ src/keys.js
@@ -1,6 +1,5 @@
 const keys = {
-    mapboxAccessToken : 'pk.eyJ1Ijoic21jdWV0byIsImEiOiJjamhsN3NqbXYzNjc3MzZwM3N2MDR1N3F4In0.zKRL_YpIG8DNww-MzDBnuA'
-}
+  mapboxAccessToken: 'pk.eyJ1Ijoic21jdWV0byIsImEiOiJjamhsN3NqbXYzNjc3MzZwM3N2MDR1N3F4In0.zKRL_YpIG8DNww-MzDBnuA',
+};
 
-
-export default keys;  
+export default keys;
